optimizer interface/super class
Moved handling of observers and assigning the self.func into the
superclass 'solver_interface'

Renamed the internal step and solving algorithms to step_alg and
solve_alg. Calls to optimizer.step() and optimizer.solve() goes to the
superclass, which calls observers and the concrete solvers stepping and
solving algorithms.

//Only really works for strictly stepping/solving algorithms. How does
this compare to DE/CMAES/other evolutionary/generation based optimizers?